resource/requirements.txt
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.6.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.16.1
gunicorn==23.0.0
Markdown==3.9
packaging==25.0
pillow==11.3.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.15.0
